feat: adds tests for keras easy hard mining
