+++ finetuner/tuner/keras/miner.py
@@ -1,9 +1,11 @@
 from typing import Tuple
 
 import tensorflow as tf
+import torch
 
 from ..miner import get_session_pairs, get_session_triplets
 from ..miner.base import BaseClassMiner, BaseSessionMiner
+from ..miner.mining_strategies import TorchStrategicMiningHelper
 
 
 class SiameseMiner(BaseClassMiner[tf.Tensor]):
@@ -38,6 +40,54 @@ class SiameseMiner(BaseClassMiner[tf.Tensor]):
         return ind1, ind2, target
 
 
+class SiameseEasyHardMiner(BaseClassMiner[tf.Tensor]):
+    def __init__(self, pos_strategy: str = 'hard', neg_strategy: str = 'hard'):
+        self.strategic_mining_helper = TorchStrategicMiningHelper(
+            pos_strategy, neg_strategy
+        )
+
+    def mine(
+        self, labels: tf.Tensor, distances: tf.Tensor
+    ) -> Tuple[tf.Tensor, tf.Tensor, tf.Tensor]:
+        """Generate all possible pairs.
+
+        :param labels: A 1D tensor of item labels (classes)
+        :param distances: A tensor matrix of pairwise distance between each two item
+            embeddings
+
+        :return: three 1D tensors, first one holding integers of first element of
+            pair, second of the second element of pair, and third one the label (0 or
+            1) for the pair for each pair
+        """
+        assert len(distances) == len(labels)
+
+        l1, l2 = tf.expand_dims(labels, 1), tf.expand_dims(labels, 0)
+        matches = tf.cast(l1 == l2, tf.uint8)
+        diffs = 1 - matches
+        matches = tf.experimental.numpy.triu(matches, 1)
+        diffs = tf.experimental.numpy.triu(diffs)
+
+        # Apply mining strategy
+        updated_matches, updated_diffs = self.strategic_mining_helper.apply_strategy(
+            torch.Tensor(matches.numpy()),
+            torch.Tensor(diffs.numpy()),
+            torch.Tensor(distances.numpy()),
+            to_numpy=True,
+        )
+
+        matches = tf.convert_to_tensor(updated_matches, dtype=matches.dtype)
+        diffs = tf.convert_to_tensor(updated_diffs, dtype=diffs.dtype)
+
+        ind1_pos, ind2_pos = tf.unstack(tf.where(matches), axis=1)
+        ind1_neg, ind2_neg = tf.unstack(tf.where(diffs), axis=1)
+
+        ind1 = tf.concat([ind1_pos, ind1_neg], axis=0)
+        ind2 = tf.concat([ind2_pos, ind2_neg], axis=0)
+
+        target = tf.concat([tf.ones_like(ind1_pos), tf.zeros_like(ind1_neg)], axis=0)
+        return ind1, ind2, target
+
+
 class TripletMiner(BaseClassMiner[tf.Tensor]):
     def mine(
         self, labels: tf.Tensor, distances: tf.Tensor
@@ -58,6 +108,48 @@ class TripletMiner(BaseClassMiner[tf.Tensor]):
         diffs = matches ^ 1
 
         matches = tf.linalg.set_diag(matches, diagonal=tf.zeros_like(labels, tf.uint8))
+        triplets = tf.expand_dims(matches, 2) * tf.expand_dims(diffs, 1)
+
+        return tf.transpose(tf.where(triplets))
+
+
+class TripletEasyHardMiner(BaseClassMiner[tf.Tensor]):
+    def __init__(self, pos_strategy: str = 'hard', neg_strategy: str = 'hard'):
+        self.strategic_mining_helper = TorchStrategicMiningHelper(
+            pos_strategy, neg_strategy
+        )
+
+    def mine(
+        self, labels: tf.Tensor, distances: tf.Tensor
+    ) -> Tuple[tf.Tensor, tf.Tensor, tf.Tensor]:
+        """Generate all possible triplets.
+
+        :param labels: A 1D tensor of item labels (classes)
+        :param distances: A tensor matrix of pairwise distance between each two item
+            embeddings
+
+        :return: three 1D tensors, holding the anchor index, positive index and
+            negative index of each triplet, respectively
+        """
+        assert len(distances) == len(labels)
+
+        l1, l2 = tf.expand_dims(labels, 1), tf.expand_dims(labels, 0)
+        matches = tf.cast(l1 == l2, tf.uint8)
+        diffs = matches ^ 1
+
+        matches = tf.linalg.set_diag(matches, diagonal=tf.zeros_like(labels, tf.uint8))
+
+        # Apply mining strategy
+        updated_matches, updated_diffs = self.strategic_mining_helper.apply_strategy(
+            torch.Tensor(matches.numpy()),
+            torch.Tensor(diffs.numpy()),
+            torch.Tensor(distances.numpy()),
+            to_numpy=True,
+        )
+
+        matches = tf.convert_to_tensor(updated_matches, dtype=matches.dtype)
+        diffs = tf.convert_to_tensor(updated_diffs, dtype=diffs.dtype)
+
         triplets = tf.expand_dims(matches, 2) * tf.expand_dims(diffs, 1)
 
         return tf.transpose(tf.where(triplets))
